+++ modernrpc/introspection.py
@@ -14,10 +14,13 @@ from django.utils.functional import cached_property
 from modernrpc.conf import settings
 
 # Define regular expressions used to parse docstring
-PARAM_REXP = re.compile(r"^[:@]param (\w+):\s?(.*)$", flags=re.MULTILINE)
-PARAM_TYPE_REXP = re.compile(r"^[:@]type (\w+):\s?(.*)$", flags=re.MULTILINE)
-RETURN_REXP = re.compile(r"^[:@]return:\s?(.*)$", flags=re.MULTILINE)
-RETURN_TYPE_REXP = re.compile(r"^[:@]rtype:\s?(.*)$", flags=re.MULTILINE)
+PARAM_REXP = re.compile(r"^[:@]param (\w+): ?(.*(?:\n[^:@].+)?)$", flags=re.MULTILINE)
+RETURN_REXP = re.compile(r"^[:@]return: ?(.*(?:\n[^:@].+)?)$", flags=re.MULTILINE)
+
+PARAM_TYPE_REXP = re.compile(r"^[:@]type (\w+): ?(.*)$", flags=re.MULTILINE)
+RETURN_TYPE_REXP = re.compile(r"^[:@]rtype ?: ?(.*)$", flags=re.MULTILINE)
+
+MULTI_SPACES = re.compile(r"\s+")
 
 
 class Introspector:
@@ -76,22 +79,27 @@ class DocstringParser:
         self.func = function
 
     @cached_property
-    def full_docstring(self) -> str:
+    def raw_docstring(self) -> str:
         """Return full content of function docstring, as extracted by inspect.getdoc()"""
         return inspect.getdoc(self.func) or ""
 
+    @staticmethod
+    def cleanup_spaces(subject: str) -> str:
+        """Basically, replaces any number of consecutive space chars (space, newline, tab, etc.) to a single space"""
+        return MULTI_SPACES.sub(" ", subject)
+
     @cached_property
-    def raw_docstring(self) -> str:
+    def text_documentation(self) -> str:
         """Return the text part of the docstring block, excluding legacy typehints and parameters and return docs"""
-        content = self.full_docstring
+        content = self.raw_docstring
         for pattern in [PARAM_REXP, PARAM_TYPE_REXP, RETURN_REXP, RETURN_TYPE_REXP]:
             content = pattern.sub("", content)
         return content.strip()
 
     @cached_property
-    def html_doc(self) -> str:
+    def html_documentation(self) -> str:
         """Convert the text part of the docstring to an HTML representation, using the parser set in settings"""
-        if not self.raw_docstring:
+        if not self.text_documentation:
             return ""
 
         if settings.MODERNRPC_DOC_FORMAT.lower() in (
@@ -101,19 +109,19 @@ class DocstringParser:
         ):
             from docutils.core import publish_parts
 
-            return publish_parts(self.raw_docstring, writer_name="html")["body"]
+            return publish_parts(self.text_documentation, writer_name="html")["body"]
 
         if settings.MODERNRPC_DOC_FORMAT.lower() in ("md", "markdown"):
             import markdown
 
-            return markdown.markdown(self.raw_docstring)
+            return markdown.markdown(self.text_documentation)
 
-        html_content = self.raw_docstring.replace("\n\n", "</p><p>").replace("\n", " ")
+        html_content = self.text_documentation.replace("\n\n", "</p><p>").replace("\n", " ")
         return f"<p>{html_content}</p>"
 
     @cached_property
     def args_doc(self) -> dict[str, str]:
-        """Return a dict with argument name as key and documentation as value. The dict will contain only
+        """Return a dict with argument name as key and documentation text as value. The dict will contain only
         documented arguments.
         Basically this method parse and extract reST documented arguments:
 
@@ -123,7 +131,7 @@ class DocstringParser:
 
             @param <argname>: Documentation for <argname>
         """
-        return dict(PARAM_REXP.findall(self.full_docstring))
+        return {param: self.cleanup_spaces(text) for param, text in PARAM_REXP.findall(self.raw_docstring)}
 
     @cached_property
     def args_types(self) -> dict[str, str]:
@@ -137,7 +145,7 @@ class DocstringParser:
 
             @type <argname>: int or str
         """
-        return dict(PARAM_TYPE_REXP.findall(self.full_docstring))
+        return dict(PARAM_TYPE_REXP.findall(self.raw_docstring))
 
     @cached_property
     def return_doc(self) -> str:
@@ -151,8 +159,8 @@ class DocstringParser:
 
         Return an empty string if return documentation is not found
         """
-        res = RETURN_REXP.search(self.full_docstring)
-        return res.group(1) if res else ""
+        res = RETURN_REXP.search(self.raw_docstring)
+        return self.cleanup_spaces(res.group(1)) if res else ""
 
     @cached_property
     def return_type(self) -> str:
@@ -166,5 +174,5 @@ class DocstringParser:
 
         Return an empty string if return type information is not found
         """
-        res = RETURN_TYPE_REXP.search(self.full_docstring)
+        res = RETURN_TYPE_REXP.search(self.raw_docstring)
         return res.group(1) if res else ""